\name{bbgdm-package}
\alias{bbgdm-package}
\docType{package}
\title{
Bayesian Bootstrap Generalised Dissimilarity Models.
}
\description{
Logitic regression model that runs a Bayesian Bootstrap
    implementation of GDM.
}
\details{
\tabular{ll}{
Package: \tab bbgdm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-19\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Skipton Woolley

Maintainer: Skipton Woolley <swoolley@museum.vic.gov.au>
}
