% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bb_gdm_check.R
\name{bbgdm.check}
\alias{bbgdm.check}
\title{Function that plots diagnostic plots from bbgdm object}
\usage{
bbgdm.check(object, plots.mfrow = c(2, 2), ...)
}
\arguments{
\item{object}{bbgdm model output}

\item{rl.col}{colour of lines}

\item{family}{logical "binomial" or "beta_dist"}
}
\value{
distances Five column matrix of sites pairs (cols 1:4) and distance as col 5.
}
\description{
Function that plots diagnostic plots from bbgdm object
}
\examples{
x <- matrix(rbinom(20*10,1,.6),20,10)# presence absence matrix
y <- simulate_covariates(x,2)
form <- ~ 1 + covar_1 + covar_2
test.bbgdm <- bbgdm(form,sp.dat=x, env.dat=y,family="binomial", dism_metric="number_non_shared",nboot=10,geo=FALSE, scale_covar=FALSE)
bbgdm.check(test.bbgdm)
}

