% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gdm_wald_test.R
\name{bbgdm.wald.test}
\alias{bbgdm.wald.test}
\title{Function to perform a non-parametric wald-test on bootstrap parameter estimates}
\usage{
bbgdm.wald.test(object, H0 = 0)
}
\arguments{
\item{object}{Returned mode from \link[bbgdm]{gdm.bb}.}
}
\value{
A table of Wald-Test statistics
}
\description{
Function to perform a non-parametric wald-test on bootstrap parameter estimates
}

