% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_geo_dist.R
\name{calc_geo_dist}
\alias{calc_geo_dist}
\title{Function that calculates geographic distances.}
\usage{
calc_geo_dist(coords, geo.type = c("euclidean", "greater_circle",
  "least_cost"), lc_data = NULL, minr = 0, maxr = NULL)
}
\arguments{
\item{coords}{A set of lats and lons for sites}

\item{geo.type}{geo.type of geographic distance to estimate, 'euclidean', 'greater circle' and 'least cost pathway'.}

\item{lc_data}{is a matrix of lons as rows and lats as cols, with a values of gradient as z values (eg,depth,elevation) or a raster of depth, elevation ect.}

\item{minr}{The minimum range to set least cost pathway, defaul is 0 to max of range.}

\item{maxr}{The maximum range to set least cost pathway, defaul is NULL to max of range.}
}
\value{
distances Five column matrix of sites pairs (cols 1:4) and distance as col 5.
}
\description{
Functions for estimating geographic distances.
}

