% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.prep-class.R
\name{table2pam}
\alias{pam2dissim}
\alias{simulate_covariates}
\alias{table2pam}
\title{data.prep objects}
\usage{
table2pam(x, site.id = "site.id", sp.id = "sp.id", abund = FALSE,
  abund.col = " No.of.specimens", siteXsp = TRUE)

pam2dissim(x, dism_metric = "number_non_shared")

simulate_covariates(x, N)
}
\arguments{
\item{x}{data to be imported into R. A table of species occurrences and site records.}

\item{site.id}{Name of column that contains site names /IDs.}

\item{sp.id}{Name of column that contains species information/ID.}

\item{abund}{LOGICAL if TRUE calculate abundances.}

\item{abund.col}{Name of column that contains abundance/counts data.}

\item{siteXsp}{LOGICAL if TRUE Returns sites as rows, sites as columns, if FALSE returns species as rows and sites as cols.}

\item{dism_metric}{If "bray_curtis" calculates Bray-Curtis dissimilarity, if "number_non_shared " returns number of shared species at site_ij which can be used in a binomial model for GDM.}

\item{x}{Presence Absence Matrix.}
}
\value{
A species by sites matrix for use in dissimilarity calculation

dism a vector of dissimilarities or count of shared species calcualted for the upper-triangle of a dissimilarity matrix.
}
\examples{
x <- matrix(rbinom(16,1,.6),4,4) #toy presence absence matrix
covr <- simulate_covariates(x,6)
}

