% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-class.R
\name{diagnostics}
\alias{bbgdm.wald.test}
\alias{diagnostics}
\alias{ggplot.bbgdm}
\alias{plot.diagnostics}
\title{diagnostics objects}
\usage{
diagnostics(object)

plot.diagnostics(object, ...)

bbgdm.wald.test(object, H0 = 0, gdm = FALSE)

\method{ggplot}{bbgdm}(object, pars = NULL, pars_labels = NULL,
  horizontal = TRUE, zero_line = TRUE, xlims = NULL, ...)
}
\arguments{
\item{object}{bbgdm model output}

\item{H0}{A numeric value giving the null hypothesis for the test. Generally zero.}

\item{gdm}{Logic if true calculates the Wald-test using variance-covariance matrix derived from the hessian matrix. Note: These estimates are probably wrong due to hessian matrix being calulated with respect to the likelihoods.}

\item{pars}{name of parameters to plot.}

\item{pars_labels}{vector of parameter labels.}

\item{horizontal}{logical. Whether or not you would like the lines to be}

\item{zero_line}{logical. Whether or not to the zero #' medians.}

\item{xlims}{custom xlims}

\item{object}{Returned model from \link[bbgdm]{bbgdm}.}
}
\description{
extracts diagnostics from \code{bbgdm} object, like random quantile residuals and plots them.
diagnostics can also call the 'Wald-like' test on a \code{bbgdm} object and returns a table of values.
}
\examples{
# fit bbgdm and extract residuals.
x <- matrix(rbinom(20*10,1,.6),20,10)# presence absence matrix
y <- simulate_covariates(x,2)
form <- ~ 1 + covar_1 + covar_2
test.bbgdm <- bbgdm(form,sp.dat=x, env.dat=y,family="binomial",
                    dism_metric="number_non_shared",nboot=10,geo=FALSE)
resids <- diagnostics(test.bbgdm)

# plot residuals
par(mfrow=c(2,2))
plot(resids)

# Undertake a wald test on bbdgm object
bbgdm.wald.test(test.bbgdm)

# plot covariate catepillar plot
library(ggplot2)
ggplot(test.bbgdm)

}

