% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diff_table_cpp}
\alias{diff_table_cpp}
\title{difference table cpp format}
\usage{
diff_table_cpp(env_dat)
}
\arguments{
\item{env_dat}{matrix  numeric matrix of site covariates}
}
\value{
diff_table absolute differences in site-pair covariates.
}
\description{
difference table cpp format
}

