% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdm_control.R
\name{gdm_control}
\alias{gdm_control}
\title{Control Parameters for Binomial Logistic Regression
Various parameters that control fitting of binomial regression models using BBgdm.}
\usage{
gdm_control(method = "BFGS", maxit = 1000, hessian = FALSE,
  trace = FALSE, start = NULL, fsmaxit = 20, fstol = 1e-05, ...)
}
\arguments{
\item{method}{characters string specifying the method argument passed to optim.}

\item{maxit}{integer specifying the maxit argument (maximal number of iterations) passed to optim.}

\item{hessian}{logical. Should the numerical Hessian matrix from the optim output be used for estimation of the covariance matrix? By default the analytical solution is employed. For details see below.}

\item{trace}{logical or integer controlling whether tracing information on the progress of the optimization should be produced (passed to optim).}

\item{start}{an optional vector with starting values for all parameters.}

\item{fsmaxit}{integer specifying maximal number of additional (quasi) Fisher scoring iterations. For details see below.}

\item{fstol}{numeric tolerance for convergence in (quasi) Fisher scoring. For details see \code{\link[stats]{optim}}.}

\item{...}{arguments passed to optim.}
}
\description{
Control Parameters for Binomial Logistic Regression
Various parameters that control fitting of binomial regression models using BBgdm.
}

