% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot.bbgdm.R
\name{ggplot.bbgdm}
\alias{ggplot.bbgdm}
\title{bbgdm catepillar plots function using ggplot}
\usage{
\method{ggplot}{bbgdm}(object, pars = NULL, pars_labels = NULL,
  horizontal = TRUE, zero_line = TRUE, xlims = NULL, ...)
}
\arguments{
\item{object}{a \code{bbgdm} object}

\item{pars}{name of parameters to plot.}

\item{pars_labels}{vector of parameter labels.}

\item{horizontal}{logical. Whether or not you would like the lines to be}

\item{zero_line}{logical. Whether or not to the zero #' medians.}

\item{xlims}{custom xlims}

\item{...}{other arguments}
}
\description{
bbgdm catepillar plots function using ggplot
}
\examples{
x <-matrix(rbinom(1:100,1,.6),10,10)# presence absence matrix
y <- simulate_covariates(x,2)
form <- ~ 1 + covar_1 + covar_2
test.bbgdm <- bbgdm(form, sp.dat=x, env.dat=y,family="binomial",
                   geo=FALSE,dism_metric="number_non_shared", nboot=10)
library(ggplot2)
ggplot(test.bbgdm)
}

