% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradfun.R
\name{GradFun}
\alias{GradFun}
\title{gradient function}
\usage{
GradFun(params, X, y, wt, offset, link)
}
\arguments{
\item{params}{initial values see start \link[stats]{optim}}

\item{X}{model matrix}

\item{y}{response variable}

\item{wt}{weights}

\item{offset}{offset vector (currently not used)}

\item{link}{link function}
}
\description{
gradient function
}

