% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invexp.R
\name{invexp}
\alias{invexp}
\title{Function to create a invexp link function for model.
For use in bbGDM}
\usage{
invexp()
}
\description{
Function to create a invexp link function for model.
For use in bbGDM
}
\examples{
inex <- invexp()
inex$linkfun(inex$linkinv(.5))  ## check invertibility
library("numDeriv")
all.equal(grad(inex$linkinv,.2),inex$mu.eta(.2))  ## check derivative
set.seed(101)
n <- 100
y <- sort(runif(n,0.1,1))
x <- sort(runif(n,1,10))
plot(-1,ylab='dissimilarity',xlab='distance',ylim=c(0,1),xlim=c(0,100),xaxt = "n")
lines(sort(predict(glm(y~x,family=binomial(link=inex)),type="response")))
}

