% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ispline.R
\name{ispline}
\alias{ispline}
\title{Function to calculate isplines.}
\usage{
ispline(x, spline.knots = 2, knots = NULL, spline.degree = 3)
}
\arguments{
\item{x}{data.frame of covariates to use as predictors in bbGDM.}

\item{spline.knots}{number knots.}

\item{knots}{NULL degrees of freedom; one can specify df rather than knots. Default = 3.}

\item{spline.degree}{degrees of freedom for ispline.}
}
\value{
A matrix of dimension c(length(x), df), where df was supplied.
}
\description{
Function to calculate isplines.
}

