% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/least_cost_dist.R
\name{least.cost.dist}
\alias{least.cost.dist}
\title{Function that calculates least cost pathway.}
\usage{
least.cost.dist(trans.mat,sites)
}
\arguments{
\item{trans.rast}{transition object as computed by trans.mat}

\item{sites}{A two-columns matrix or data.frame containing latitude and longitude for 2 or more locations.}
}
\value{
distance matrix in meters between all possible pairs of locations
}
\description{
Function that calculates least cost pathway.
}
\details{
least.cost.dist computes least cost distances between 2 or more locations. This function relies on the package gdistance (van Etten, 2011. http://CRAN.R-project.org/package=gdistance) and on the trans.mat function to define a range of depths where the paths are possible.
}

