% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_glm_control_nlminb.R
\name{logit_glm_control_nlminb}
\alias{logit_glm_control_nlminb}
\title{Control Parameters for Binomial Logistic Regression
Various parameters that control fitting of binomial regression models using BBgdm.}
\usage{
logit_glm_control_nlminb(inter.max = 1000, hessian = NULL, trace = 0,
  start = NULL, eval.max = 200, ...)
}
\arguments{
\item{inter.max}{integer specifying the inter.max argument (maximal number of iterations) passed to optim.}

\item{hessian}{logical. Should the numerical Hessian matrix from the optim output be used for estimation of the covariance matrix? By default the analytical solution is employed. For details see below.}

\item{trace}{logical or integer controlling whether tracing information on the progress of the optimization should be produced (passed to optim).}

\item{start}{an optional vector with starting values for all parameters.}

\item{...}{arguments passed to optim.}

\item{method}{characters string specifying the method argument passed to optim.}

\item{fsmaxit}{integer specifying maximal number of additional (quasi) Fisher scoring iterations. For details see below.}

\item{fstol}{numeric tolerance for convergence in (quasi) Fisher scoring. For details see \link[stats]{optim}.}
}
\description{
Control Parameters for Binomial Logistic Regression
Various parameters that control fitting of binomial regression models using BBgdm.
}

