% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logit_binomial_glm_fit.R
\name{logit_glm_fit}
\alias{logit_glm_fit}
\title{Function to perform postive binomial logistic regression.}
\usage{
logit_glm_fit(X, y, wt = NULL, offset, optim.meth = TRUE, est.var = TRUE,
  trace = FALSE, control = logit_glm_control(...), ...)
}
\arguments{
\item{X}{Model matirx of predictors, see \link[stats]{model.matrix}.}

\item{y}{Model response, see \link[stats]{model.response}.}

\item{wt}{weights for model.}

\item{control}{control option from optim see \link[BayesbootGDM]{logit_glm_control} or \link[stats]{optim}}

\item{scale.covar}{logical If TRUE centre predictor variables.}
}
\value{
fit fitted logistic binomial model as per optim methods
}
\description{
Function to perform postive binomial logistic regression.
}

