% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikfun.R
\name{LogLikFun}
\alias{LogLikFun}
\title{log-likelihood function}
\usage{
LogLikFun(params, X, y, wt, offset, link)
}
\arguments{
\item{params}{initial values see start \link[stats]{optim}}

\item{X}{model matrix}

\item{y}{response variable}

\item{wt}{weights}

\item{offset}{offset vector (currently not used)}

\item{link}{link function}
}
\description{
log-likelihood function
}

