% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_exp_link.R
\name{negexp}
\alias{negexp}
\title{Negative exponential link function.}
\usage{
negexp()
}
\value{
negexp() link function
}
\description{
A object of class \code{\link[stats]{glm}}, a list with components
}

