% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Function to normalise isplines. 
prediction spline transform
Transform spline data based on parameters}
\usage{
normalise(x, standardize = "zscore")
}
\arguments{
\item{x}{spline data}

\item{standardize}{'zscore'}
}
\description{
Function to normalise isplines. 
prediction spline transform
Transform spline data based on parameters
}

