% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam2dissim.R
\name{pam2dissim}
\alias{pam2dissim}
\title{Calculate Dissimilarities from Presence Absence Matrix}
\usage{
pam2dissim(x, dism_metric = "number_non_shared")
}
\arguments{
\item{x}{Presence Absence Matrix.}

\item{dism_metric}{If "bray_curtis" calculates Bray-Curtis dissimilarity, if "number_non_shared " returns number of shared species at site_ij which can be used in a binomial model for GDM.}
}
\value{
dism a vector of dissimilarities or count of shared species calcualted for the upper-triangle of a dissimilarity matrix.
}
\description{
Creates a vector of dissimilarities from presence absence matrix.
}

