% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bbgdm_catepillar_plot.R
\name{plot.bbgdm.caterpillar}
\alias{plot.bbgdm.caterpillar}
\title{bbgdm catepillar plots function}
\usage{
\method{plot}{bbgdm.caterpillar}(object, pars = NULL, pars_labels = NULL,
  horizontal = TRUE, zero_line = TRUE, xlims = NULL)
}
\arguments{
\item{object}{a \code{bbgdm} object}

\item{pars}{name of parameters to plot.}

\item{pars_labels}{vector of parameter labels.}

\item{horizontal}{logical. Whether or not you would like the lines to be
horizontal}

\item{zero_line}{logical. Whether or not to the zero #' medians.}
}
\description{
bbgdm catepillar plots function
}

