% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_gdm_bb.R
\name{plot.bbgdm}
\alias{plot.bbgdm}
\title{Function for plotting GDM w/ Bayesian Bootstrap.}
\usage{
plot.bbgdm(object, plot.layout = c(1, 1), plot.colour = "black",
  plot.linewidth = 2, line.col = "red", ...)
}
\arguments{
\item{object}{As derived from bbgdm function}

\item{plot.layout}{par settings for plot. default is c(2,2) (four plots).}

\item{plot.colour}{colour of plot points}

\item{plot.linewidth}{line weight}

\item{line.col}{line colour}

\item{...}{other plot arguments}
}
\value{
a plots from bbgdm
}
\description{
Plots a Generalised dissimilarity model with bayesian bootstrap.
}
\examples{
x <- matrix(rbinom(20*10,1,.6),20,10)# presence absence matrix
y <- simulate_covariates(x,2)
form <- ~ 1 + covar_1 + covar_2
test.bbgdm <- bbgdm(form,sp.dat=x, env.dat=y,family="binomial", dism_metric="number_non_shared",
                   nboot=10,geo=FALSE)
plot(test.bbgdm)
}

