% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_gdm_bb.R
\name{plot.gdm.bb}
\alias{plot.gdm.bb}
\title{Function for plotting GDM w/ Bayesian Bootstrap.}
\usage{
\method{plot}{gdm.bb}(object, plot.layout = c(2, 2), plot.colour = "black",
  plot.linewidth = 2, line.col = "red")
}
\arguments{
\item{plot.layout}{par settings for plot. default is c(2,2) (four plots).}

\item{gdm.bb_object}{As derived from gdm.bb function}
}
\value{
a plots from gdm.bb
}
\description{
Plots a Generalised dissimilarity model with bayesian bootstrap.
}
\examples{
x <- matrix(rbinom(20*10,1,.6),20,10)# presence absence matrix
y <- simulate_covariates(x,2)
form <- ~ 1 + covar_1 + covar_2
test.gdm.bb <- gdm.bb(form,sp.dat=x, env.dat=y,family="binomial", dism_metric="number_shared",nboot=10, scale_covar=T)
plot(test.gdm.bb)
}

