% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_response_sim_gdm_bb.R
\name{plotResponseSim}
\alias{plotResponseSim}
\title{Function for plotting bb.GDM response curves.}
\usage{
plotResponseSim(object, plotdim = c(1, 1))
}
\arguments{
\item{gdm.bb_object}{As derived from gdm.bb function}

\item{plot.layout}{par settings for plot. default is c(2,2) (four plots).}
}
\value{
a plots from gdm.bb
}
\description{
Plots a Generalised dissimilarity model with bayesian bootstrap.
}
\examples{
library(bbgdm)
theta <- matrix(c(.5,2),1,2)
form <- y~1+x
nsp <- 250
Nsite <- 20
ndissim <- (Nsite*(Nsite-1))/2
dat <- data.frame(y=rep(1,ndissim),x=seq(0,4,length.out = ndissim))#imaginary difference in a covariate
sim1 <- simulate_dism_data(form,dat, matrix(c(qlogis(.25),1),1,2),nsp)
y <- data.frame(SharedSpp_ij=sim1$nnss, MaxSpp_ij=sim1$max_sp_site)
test1 <- sim_bbgdm(X=sim1$X,y=y,Nsite = Nsite,nboots = 100)
plotResponseSim(test1)
}

