% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pred_bbgdm_cpp}
\alias{pred_bbgdm_cpp}
\title{predict bbgdm cpp function}
\usage{
pred_bbgdm_cpp(raster_stack, output_raster, neighbourhood, bbgdm_coef)
}
\arguments{
\item{raster_stack}{raster stack of i-spline transformed environmental layers.}

\item{output_raster}{binary raster one for cells to predict to and NA for areas not to predict too.}

\item{neighbourhood}{numeric matrix which creates the neighbourhood average.}

\item{bbgdm_coef}{I-spline coefficents from bbgdm model fit.}
}
\description{
predict bbgdm cpp function
}

