% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bbgdm.R
\name{predict.bbgdm}
\alias{predict.bbgdm}
\title{Spatiall predict a bayesian bootstrap generalised dissimilarity model.}
\usage{
predict.bbgdm(object, data, neighbourhood = NULL, outer = FALSE,
  uncertainty = TRUE, ...)
}
\arguments{
\item{object}{As derived from bbgdm function}

\item{data}{raster stack of the same covariates used to fit model object for the region you wish to predict too.}

\item{outer}{logical default is FALSE, if TRUE only calculates the outer edge of neighbourhoods area.}

\item{uncertainty}{logical if TRUE predict will return a list with two rasters the mean estimate and the uncertainty (defined as the coefficent of variation)}

\item{neigbourhood}{int default is three, number of neighbouring cells to estimate mean dissimilarity.}
}
\value{
raster of mean turnover estimated based on neighbourhood distance.
}
\description{
Spatiall predict a bayesian bootstrap generalised dissimilarity model.
}

