% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict_gdm_bb.R
\name{predict.gdm.bb}
\alias{predict.gdm.bb}
\title{Function for plotting GDM w/ Bayesian Bootstrap.}
\usage{
\method{predict}{gdm.bb}(object, newdata = NULL, type = c("link",
  "response", "terms"), se.fit = FALSE, dispersion = NULL, terms = NULL,
  na.action = na.pass, ...)
}
\arguments{
\item{type}{the type of prediction required. The default is on the scale of the linear predictors; the alternative "response" is on the scale of the response variable. Thus for a default binomial model the default predictions are of log-odds (probabilities on logit scale) and type = "response" gives the predicted probabilities. The "terms" option returns a matrix giving the fitted values of each term in the model formula on the linear predictor scale.   The value of this argument can be abbreviated.}

\item{se.fit}{logical switch indicating if standard errors are required.}

\item{dispersion}{the dispersion of the GLM fit to be assumed in computing the standard errors. If omitted, that returned by summary applied to the object is used.}

\item{terms}{with type = "terms" by default all terms are returned. A character vector specifies which terms are to be returned}

\item{na.action}{function determining what should be done with missing values in newdata. The default is to predict NA.}

\item{gdm.bb_object}{As derived from gdm.bb function}

\item{preddata}{new environmental and spatial covariate data to predict GDM.bb model to.}
}
\description{
Plots a Generalised dissimilarity model with bayesian bootstrap.
}
\examples{
x <- matrix(rbinom(1:100,1,.6),10,10)# presence absence matrix
form <- ~ 1 + covar_1 + covar_2
test.gdm.bb <- gdm.bb(form,family="binomial", dism_metric="number_shared", nboot=200, x, sim_covar=TRUE,scale_covar=FALSE)
predict(test.gdm.bb,preddata)
}

