% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_spline_trans.R
\name{predict.spline.trans}
\alias{predict.spline.trans}
\title{Predict splines transformations}
\usage{
\method{predict}{spline.trans}(x, attrib = NULL, values = NULL,
  standardization = NULL, splineInterval = NULL, splineDegree = NULL)
}
\arguments{
\item{x}{list of predictor values}

\item{attrib}{Spline parameters from model.}

\item{values}{NULL}

\item{standardization}{NULL}

\item{splineInterval}{NULL}

\item{splineDegree}{NULL}
}
\value{
I Spline predictions
}
\description{
Predict splines transformations
}

