% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_bbgdm.R
\name{sim_bbgdm}
\alias{sim_bbgdm}
\title{Simulate bbgdms}
\usage{
sim_bbgdm(X, y, offset = NULL, Nsite, nboots = 10,
  control = logit_glm_control(...), ...)
}
\arguments{
\item{X}{model matrix with intercept = 1}

\item{y}{no-non-shared-sp as two column response fro binomial model}

\item{Nsite}{No orginal sites}

\item{nboots}{number of bootstraps}
}
\value{
a gdm.bb model object
}
\description{
Simulate bbgdms
}
\examples{
library(bbgdm)
theta <- matrix(c(.5,2),1,2)
form <- y~1+x
nsp <- 250
Nsite <- 50
ndissim <- (Nsite*(Nsite-1))/2
dat <- data.frame(y=rep(1,ndissim),x=seq(0,4,length.out = ndissim))#imaginary difference in a covariate
sim1 <- simulate_dism_data(form,dat, matrix(c(qlogis(.25),.1),1,2),nsp)
y <- data.frame(SharedSpp_ij=sim1$nnss, MaxSpp_ij=sim1$max_sp_site)
test1 <- sim_bbgdm(sim1$X,y,Nsite,10)
}

