% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_covariates.R
\name{simulate_covariates}
\alias{simulate_covariates}
\title{Simulate Covariates for bbGDM}
\usage{
simulate_covariates(x, N)
}
\arguments{
\item{x}{Presence Absence Matrix.}

\item{N}{number of covariates to simulate}
}
\value{
beta N covariates for number of sites of x.
}
\description{
Creates N covarites for testing bbGDM.
}
\examples{
x <- matrix(rbinom(16,1,.6),4,4) #toy presence absence matrix
covr <- simulate_covariates(x,6)
}

