% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dism_data.R
\name{simulate_dism_data}
\alias{simulate_dism_data}
\title{Simulate number of non-shared species across a gradient.}
\usage{
simulate_dism_data(form, dat, theta, nsp)
}
\arguments{
\item{form}{formula for simulation.}

\item{dat}{data to simulate dissimilarites.}

\item{theta}{response parameters.}

\item{nsp}{Number of species underlying simulation.}
}
\value{
list containing, form,dat,theta,nsp,variance and max_sp_site.
}
\description{
Simulate number of non-shared species across a gradient.
}
\examples{
library(bbgdm)
theta <- matrix(c(.5,2),1,2)
form <- y~1+x
nsp <- 250
Nsite <- 50
ndissim <- (Nsite*(Nsite-1))/2
dat <- data.frame(y=rep(1,ndissim),x=seq(0,4,length.out = ndissim))#imaginary difference in a covariate
sim1 <- simulate_dism_data(form,dat, matrix(c(qlogis(.25),.1),1,2),nsp)
plot(sim1$X[,2],sim1$dism,ylim=c(0,1),pch=16,cex=.75)
}

