% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spline_transform.R
\name{spline.trans}
\alias{spline.trans}
\title{Transformation of Covariates to Splines.}
\usage{
spline.trans(x, spline_type = "bspline", spline_df = 2,
  spline_knots = NULL)
}
\arguments{
\item{x}{data.frame of covariates to use as predictors in bbGDM.}

\item{spline_type}{If "bspline" calculates bs spline from spline package. If "ispline" calculates ispline.}

\item{spline_df}{degrees of freedom; one can specify df rather than knots. Default = 3.}

\item{spline_knots}{The internal breakpoints that define the spline. Default = NULL for bs spline and 1 for ispline.}
}
\value{
A matrix of dimension c(length(x), df), where df was supplied.
}
\description{
Splines to be used in bbGDM; bsplines and isplines can be calculated.
}

