% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_trans_spline.R
\name{spline_trans_for_pred}
\alias{spline_trans_for_pred}
\title{prediction spline transform
Transform spline data based on parameters}
\usage{
spline_trans_for_pred(x, attrib = NULL, values = NULL,
  standardization = NULL, splineInterval = NULL, splineDegree = NULL)
}
\arguments{
\item{x}{list of predictor values}

\item{attrib}{Spline parameters from model.}

\item{values}{NULL}

\item{standardization}{NULL}

\item{splineInterval}{NULL}

\item{splineDegree}{NULL}
}
\value{
I-Spline predictions
}
\description{
prediction spline transform
Transform spline data based on parameters
}

