% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table2pam.R
\name{table2pam}
\alias{table2pam}
\title{Transform Occurrence Data to Presence Absence Matrix}
\usage{
table2pam(x, site.id = "site.id", sp.id = "sp.id", abund = FALSE,
  abund.col = " No.of.specimens", siteXsp = TRUE)
}
\arguments{
\item{x}{data to be imported into R. A table of species occurrences and site records.}

\item{site.id}{Name of column that contains site names /IDs.}

\item{sp.id}{Name of column that contains species information/ID.}

\item{abund}{LOGICAL if TRUE calculate abundances.}

\item{abund.col}{Name of column that contains abundance/counts data.}

\item{siteXsp}{LOGICAL if TRUE Returns sites as rows, sites as columns, if FALSE returns species as rows and sites as cols.}
}
\value{
A species by sites matrix for use in dissimilarity calculation
}
\description{
Creates a pa matrix from occurrence data for GDM models.
}

