% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_rast_fun.R
\name{trans_rast_fun}
\alias{trans_rast_fun}
\title{Create transition raster}
\usage{
trans_rast_fun(data,min.range=0,max.range=NULL)
}
\arguments{
\item{data}{A matrix of lons as rows, lats as cols and value as gradient (eg depth or elevation), or a raster}

\item{min.range}{Numeric. The range of depth between which the path will be possible. The default (min.depth=0 and max.depth=NULL) indicates that the transition between cells of the grid is possible between 0 meters depth and the maximum range.}

\item{max.range}{Numeric. The range of depth between which the path will be possible. The default (min.depth=0 and max.depth=NULL) indicates that the transition between cells of the grid is possible between 0 meters depth and the maximum range.}
}
\value{
A transition raster for least_cost_dist function.
}
\description{
Creates a transition raster from matrix or raster to be used to compute least cost distances between sites.
}
\note{
This function is slow, use courser resolution lons x lat or raster grids to speed up.
}

