% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbgdm.fit-class.R
\name{bbgdm.fit}
\alias{bbgdm.control}
\alias{bbgdm.fit}
\alias{negexp}
\title{bbgdm.fit objects}
\usage{
bbgdm.fit(X, y, wt = NULL, link, optim.meth = "optim", est.var = TRUE,
  trace = FALSE, prior = FALSE, control = bbgdm.control(...))

bbgdm.control(method = "BFGS", maxit = 1000, hessian = FALSE,
  trace = FALSE, start = NULL, fsmaxit = 20, fstol = 1e-05, ...)

negexp()
}
\arguments{
\item{X}{Model matirx of predictors, see \code{\link[stats]{model.matrix}}.}

\item{y}{Model response, see \code{\link[stats]{model.response}}.}

\item{wt}{weights for model.}

\item{link}{character link functions. default is 'logit', can call other
link functions \code{\link[stats]{make.link}} or a 'negexp' custom link function (e.g., Ferrier etal., 2007).}

\item{optim.meth}{optimisation method options avaliable are 'optim' and 'nlmnib',
calls either method \code{\link[stats]{optim}} or \code{\link[stats]{nlminb}}.}

\item{est.var}{logical if true estimated parameter variance using optimiser.}

\item{trace}{options looks at \code{\link[stats]{optim}} for details.}

\item{prior}{numeric vector of starting values for intercept and splines.}

\item{control}{control option from optim see \code{\link[bbgdm]{bbgdm.control}} or \code{\link[stats]{optim}}.}

\item{method}{characters string specifying the method argument passed to optim.}

\item{maxit}{integer specifying the maxit argument (maximal number of iterations) passed to optim.}

\item{hessian}{logical. Should the numerical Hessian matrix from the optim output be used for estimation of the covariance matrix? By default the analytical solution is employed. For details see below.}

\item{start}{an optional vector with starting values for all parameters.}

\item{fsmaxit}{integer specifying maximal number of additional (quasi) Fisher scoring iterations. For details see below.}

\item{fstol}{numeric tolerance for convergence in (quasi) Fisher scoring. For details see \code{\link[stats]{optim}}.}

\item{\dots}{for \code{bbgdm.fit()} or more \code{bbgdm.control()}.}
}
\value{
fit single gdm
}
\description{
function for fitting a \code{bbgdm} model. This is essential a single gdm
and builds the core of \code{bbgdm}. Like \code{\link[stats]{glm.fit}} a number of options
can be defined. Like: optimisation method \code{\link[stats]{optim}} or \code{\link[stats]{nlminb}},
link function \code{\link[stats]{make.link}}.
}
\author{
Skipton Woolley
}

