% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbgdm-class.R, R/bbgdm-package.R
\docType{package}
\name{bbgdm}
\alias{bbgdm}
\alias{bbgdm-package}
\alias{plot.bbgdm}
\alias{predict.bbgdm}
\title{bbgdm objects}
\usage{
bbgdm(form, sp.dat, env.dat, family = "binomial", link = "logit",
  dism_metric = "number_non_shared", nboot = 100, spline_type = "ispline",
  spline_df = 2, spline_knots = 1, geo = TRUE, geo.type = "euclidean",
  coord.names = c("X", "Y"), optim.meth = "nlmnib", est.var = FALSE,
  trace = FALSE, prior = FALSE, control = bbgdm.control())

plot.bbgdm(object, ...)

predict.bbgdm(object, data, neighbourhood = NULL, outer = FALSE,
  uncertainty = TRUE, ...)
}
\arguments{
\item{form}{formula for bbgdm model}

\item{sp.dat}{presence absence matrix, sp as columns sites as rows.}

\item{env.dat}{environmental or spatial covariates at each site.}

\item{family}{a description of the error distribution and link function to be used in the model.
Currently "binomial" suppported. This can be a character string naming a family function,
a family function or the result of a call to a family function.}

\item{link}{a character string that assigns the link function to apply within the binomial model.
Default is 'logit', but 'negexp' and other binomial link functions can be called.}

\item{dism_metric}{dissimilarity metric to calculate for model. "bray_curtis" or "number_non_shared"
currently avaliable.}

\item{nboot}{number of Bayesian Bootstraps to run, this is used to estimate variance around GDM models.
Default is 100 iterations.}

\item{spline_type}{type of spline to use in GDM model. Default is monotonic isplines.
Options are: "ispline" or "bspline".}

\item{spline_df}{Number of spline degrees of freedom.}

\item{spline_knots}{Number of spline knots.}

\item{geo}{logical If true geographic distance is calculated if}

\item{geo.type}{type of geographic distance to estimate, can call 'euclidean' and 'greater_circle'.}

\item{coord.names}{character.vector names of coordinates, default is c("X","Y")}

\item{optim.meth}{optimisation method options avaliable are 'optim' and 'nlmnib'}

\item{est.var}{logical if true estimated parameter variance using optimiser.}

\item{trace}{logical print extra optimisation outputs}

\item{prior}{numeric vector of starting values for intercept and splines}

\item{control}{control options for gdm calls \link[bbgdm]{bbgdm.control} as default.}

\item{object}{As derived from bbgdm function}

\item{data}{raster stack of the same covariates used to fit model object for the region you wish to predict too.}

\item{neighbourhood}{int default is three, number of neighbouring cells to estimate mean dissimilarity.}

\item{outer}{logical default is FALSE, if TRUE only calculates the outer edge of neighbourhoods area.}

\item{uncertainty}{logical if TRUE predict will return a list with two rasters the mean estimate and the uncertainty (defined as the coefficent of variation)}

\item{\dots}{for \code{bbgdm()}: one or more \code{plot()} and
\code{predict()}: further arguments passed to or from other methods}
}
\value{
a bbgdm model object

raster of mean turnover estimated based on neighbourhood distance.
}
\description{
creates a \code{bbgdm} model, comprising multiple
  gdms. \code{bbgdm} models are core of \code{bbgdm}, different parameterisation can
  be achieve similar to a \code{\link[stats]{glm}}, see \code{bbgdm.fit} for more details.

bbgdm package
}
\examples{
sp.dat <- matrix(rbinom(200,1,.6),20,10)# presence absence matrix
env.dat <- simulate_covariates(sp.dat,2)
form <- ~ 1 + covar_1 + covar_2
test.bbgdm <- bbgdm(form,sp.dat, env.dat,family="binomial",dism_metric="number_non_shared",
                    nboot=10, geo=FALSE,optim.meth='nlmnib')
#plot bbgdm fit
plot(test.bbgdm)

}

