% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdm_wald_test.R
\name{bbgdm.wald.test}
\alias{bbgdm.wald.test}
\title{Function to perform a non-parametric wald-test on bootstrap parameter estimates}
\usage{
bbgdm.wald.test(object, H0 = 0, gdm = FALSE)
}
\arguments{
\item{object}{Returned model from \link[bbgdm]{bbgdm}.}

\item{H0}{A numeric value giving the null hypothesis for the test. Generally zero.}

\item{gdm}{Logic if true calculates the Wald-test using variance-covariance matrix derived from the hessian matrix. Note: These estimates are probably wrong due to hessian matrix being calulated with respect to the likelihoods.}
}
\value{
A table of Wald-Test statistics
}
\description{
Function to perform a non-parametric wald-test on bootstrap parameter estimates
}

