% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diff_table_cpp}
\alias{diff_table_cpp}
\title{diff_table cpp function}
\usage{
diff_table_cpp(env_dat)
}
\arguments{
\item{env_dat}{matrix of sites x environmental covariates.}
}
\value{
diff_table difference between covariates between site-pairs
}
\description{
diff_table cpp function
}

