% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissim_table.R
\name{dissim_table}
\alias{dissim_table}
\title{Create dissimilarity table for bbGDM.}
\usage{
dissim_table(sp.dat, env.dat, dism_metric = "bray_curtis",
  spline_type = "bspline", spline_df = 1, spline_knots = 2,
  coord.names = c("X", "Y"), geo = FALSE, geo.type = "euclidean",
  lc_data = NULL, minr = NULL, maxr = NULL)
}
\arguments{
\item{sp.dat}{Presence absence matrix.}

\item{env.dat}{Matrix of covariates for bbGDM; if using own data as covariates make sure to set sim_covar=FALSE, otherwise simulated covariates will be caluclated instead.}

\item{dism_metric}{Dissimilarity to caluclate; If "bray_curtis" calculates Bray-Curtis dissimilarity, if "number_shared " returns number of shared species at site_ij which can be used in a binomial model for GDM.}

\item{spline_type}{If "bspline" calculates bs spline from spline package. If "ispline" calculates ispline.}

\item{spline_df}{degrees of freedom; one can specify df rather than knots. Default = 3.}

\item{spline_knots}{The internal breakpoints that define the spline. Default = NULL for bs spline and 1 for ispline.}
}
\value{
diff_table dissimilarity table; creates a table of dissimilarities, and difference of covariates between each sites_ij. If select dissim="number_shared", will return "SharedSpp_ij","MaxSpp_ij", which can be used as response variables in a binomial bbGDM.
}
\description{
Calculates dissimilarity table for bbGMD from presences absence matrix (pam).
}
\examples{
x <- matrix(sample(0:1,100, replace=T),10,10) #toy presence absence matrix
y <- simulate_covariates(x,4)
diff_table <- dissim_table(x,y,dissim="bray_curtis")
}

