% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity-class.R, R/spline-class.R
\name{dissim_table}
\alias{dissim_table}
\alias{is.dissim_table}
\alias{is.spline}
\title{dissimilarity objects}
\usage{
dissim_table(sp.dat, env.dat, dism_metric = "number_non_shared",
  spline_type = "ispline", spline_df = 1, spline_knots = 2,
  coord.names = c("X", "Y"), geo = FALSE, geo.type = "euclidean")

is.dissim_table(x)

is.spline(object)
}
\arguments{
\item{sp.dat}{Presence absence matrix.}

\item{env.dat}{Matrix of covariates for bbgdm; if using own data as covariates make sure to set sim_covar=FALSE,
otherwise simulated covariates will be caluclated instead.}

\item{dism_metric}{Dissimilarity to caluclate; If "bray_curtis" calculates Bray-Curtis dissimilarity,
if "number_non_shared" returns number of shared species at site_ij which can be used in a binomial model for GDM.}

\item{spline_type}{If "ispline" calculates ispline. If "bspline" calculates bc from spline package.}

\item{spline_df}{degrees of freedom; one can specify df rather than knots. Default = 3.}

\item{spline_knots}{The internal breakpoints that define the spline. Default = NULL for bs spline and 1 for ispline.}

\item{coord.names}{c("X","Y") character.vector names of coordinates.}

\item{geo}{logical If true geographic distance is calculated if}

\item{geo.type}{type of geographic distance to estimate, can call 'euclidean','greater_circle' or 'least_cost'. If least_cost is called extra parameters are required (lc_data, minr and maxr).}

\item{x}{dissim_table class}

\item{object}{spline class object}
}
\value{
dissim_table dissimilarity table; creates a table of dissimilarities, and difference of covariates between each sites_ij.
 If select dissim="number_non_shared", will return "nonsharedspp_ij","sumspp_ij",
 which can be used as response variables in a binomial bbgdm.
}
\description{
creates a \code{dissim_table} which is required by \code{bbgdm}
}
\examples{
x <- matrix(rbinom(100,1,.5),10,10) #toy presence absence matrix
y <- simulate_covariates(x,4)
diff_table <- dissim_table(x,y,dism_metric="number_non_shared")
}

