% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gcdist.R
\name{gcdist}
\alias{gcdist}
\title{greater circle dist}
\usage{
gcdist(x1, y1, x2, y2)
}
\arguments{
\item{lon1,}{lon2, lat1, lat2 coordinates between of two sites}
}
\value{
greater circle distance.
}
\description{
Takes in any numeric value and squares itFunction for calculating greater circle dist
}

