% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdm_fit.R
\name{gdm_fit}
\alias{gdm_fit}
\title{Function to perform postive binomial logistic regression.}
\usage{
gdm_fit(X, y, wt = NULL, offset, link, optim.meth = "optim",
  est.var = TRUE, trace = FALSE, prior = FALSE,
  control = gdm_control(...), ...)
}
\arguments{
\item{X}{Model matirx of predictors, see \code{\link[stats]{model.matrix}}.}

\item{y}{Model response, see \code{\link[stats]{model.response}}.}

\item{wt}{weights for model.}

\item{offset}{offset values}

\item{link}{character link functions. default is 'logit', can call 'negexp'}

\item{optim.meth}{optimisation method options avaliable are 'optim' and 'nlmnib'}

\item{est.var}{logical if true estimated parameter variance using optimiser.}

\item{trace}{trace options looks at \code{\link[stats]{optim}} for details.}

\item{prior}{numeric vector of starting values for intercept and splines}

\item{control}{control option from optim see \code{\link[bbgdm]{gdm_control}} or \code{\link[stats]{optim}}}

\item{...}{other arguments}
}
\value{
fit fitted logistic binomial model as per optim methods
}
\description{
Function to perform postive binomial logistic regression.
}

