% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdm_fit.R
\name{gdm_fit}
\alias{gdm_fit}
\title{Function to perform postive binomial logistic regression.}
\usage{
gdm_fit(X, y, wt = NULL, offset, link, optim.meth = "optim",
  est.var = TRUE, trace = FALSE, prior = FALSE,
  control = gdm_fit_control(...), ...)
}
\arguments{
\item{X}{Model matirx of predictors, see \link[stats]{model.matrix}.}

\item{y}{Model response, see \link[stats]{model.response}.}

\item{wt}{weights for model.}

\item{control}{control option from optim see \link[BayesbootGDM]{gdm_fit_control} or \link[stats]{optim}}

\item{scale.covar}{logical If TRUE centre predictor variables.}
}
\value{
fit fitted logistic binomial model as per optim methods
}
\description{
Function to perform postive binomial logistic regression.
}

