% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/negative_exp_link.R
\name{negexp}
\alias{negexp}
\title{Negative exponentional link function.
A object of class \link[stats]{link-glm}, a list with components}
\usage{
negexp()
}
\arguments{
\item{linkfun}{Link function function(mu)}

\item{linkinv}{Inverse link function function(eta)}

\item{mu.eta}{Derivative function(eta) dmu/deta}

\item{valideta}{function(eta){ TRUE if eta is in the domain of linkinv }.}

\item{name}{a name to be used for the link}
}
\description{
Negative exponentional link function.
A object of class \link[stats]{link-glm}, a list with components
}

