% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pam2dissim.R
\name{pam2dissim}
\alias{pam2dissim}
\title{Calculate Dissimilarities from Presence Absence Matrix}
\usage{
pam2dissim(x, dissim = "bray_curtis")
}
\arguments{
\item{x}{Presence Absence Matrix.}

\item{dissim}{If "bray_curtis" calculates Bray-Curtis dissimilarity, if "number_shared " returns number of shared species at site_ij which can be used in a binomial model for GDM.}
}
\value{
dism a vector of dissimilarities or count of shared species calcualted for the upper-triangle of a dissimilarity matrix.
}
\description{
Creates a vector of dissimilarities from presence absence matrix.
}

