% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_response_gdm_bb.R
\name{plotResponse}
\alias{plotResponse}
\title{Function for plotting bbgdm response curves.}
\usage{
plotResponse(object, plotdim = c(2, 2), ...)
}
\arguments{
\item{object}{As derived from bbgdm function}

\item{plotdim}{par settings for plot. default is c(2,2) (four plots).}
}
\value{
a plots from bbgdm
}
\description{
Plots a Generalised dissimilarity model with bayesian bootstrap.
}
\examples{
x <-matrix(rbinom(1:100,1,.6),10,10)# presence absence matrix
y <- simulate_covariates(x,2)
form <- ~ 1 + covar_1 + covar_2
test.bbgdm <- bbgdm(form, sp.dat=x, env.dat=y,family="binomial",
                   geo=FALSE,dism_metric="number_non_shared", nboot=10)
plotResponse(test.bbgdm)
}

