% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_response_gdm_bb.R
\name{plotResponse}
\alias{plotResponse}
\title{Function for plotting bb.GDM response curves.}
\usage{
plotResponse(object, plotdim = c(2, 2), plot.error = F, monotonic = F,
  ...)
}
\arguments{
\item{gdm.bb_object}{As derived from gdm.bb function}

\item{plot.layout}{par settings for plot. default is c(2,2) (four plots).}
}
\value{
a plots from gdm.bb
}
\description{
Plots a Generalised dissimilarity model with bayesian bootstrap.
}
\examples{
x<-matrix(rbinom(1:100,1,.6),10,10)# presence absence matrix
#' env.dat <- simulate_covariates(x,2)
form <- y ~ covar_1 + covar_2
test.gdm.bb <- gdm.bb(form,family="binomial", dism_metric="number_shared", nboot=100)
plotResponse(test.gdm.bb,plot.error=T)
}

