% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rqr_function.R
\name{rqr_function}
\alias{rqr_function}
\title{Randomised quantile residual function
Computes randomised quantile resiudeals for bbgdm}
\usage{
rqr_function(X, y, coefs, offset, link)
}
\arguments{
\item{X}{model matrix}

\item{y}{response variable}

\item{coefs}{coefficents from bbgdm model}

\item{offset}{offset vector (currently not used)}

\item{link}{link function}
}
\value{
res randomised quantile residuals
}
\description{
Randomised quantile residual function
Computes randomised quantile resiudeals for bbgdm
}

