% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline-class.R
\name{spline_trans_for_pred}
\alias{ispline}
\alias{spline.trans}
\alias{spline_trans_for_pred}
\title{spline objects}
\usage{
spline_trans_for_pred(x, attrib = NULL, values = NULL,
  standardization = NULL, splineInterval = NULL, splineDegree = NULL)

spline.trans(x, spline_type = "bspline", spline_df = 2, spline_knots = 1)

ispline(x, spline.knots = 2, knots = NULL, spline.degree = 3)
}
\arguments{
\item{x}{list of predictor values}

\item{attrib}{Spline parameters from model.}

\item{values}{NULL}

\item{standardization}{NULL}

\item{splineInterval}{NULL}

\item{splineDegree}{NULL}

\item{spline_type}{If "bspline" calculates bs spline from spline package. If "ispline" calculates ispline.}

\item{spline_df}{degrees of freedom; one can specify df rather than knots. Default = 3.}

\item{spline_knots}{The internal breakpoints that define the spline. Default = NULL for bs spline and 1 for ispline.}

\item{spline.knots}{number knots.}

\item{knots}{NULL degrees of freedom; one can specify df rather than knots. Default = 3.}

\item{spline.degree}{degrees of freedom for ispline.}

\item{x}{data.frame of covariates to use as predictors in bbGDM.}

\item{x}{data.frame of covariates to use as predictors in bbGDM.}
}
\value{
I-Spline predictions

A matrix of dimension c(length(x), df), where df was supplied.

A matrix of dimension c(length(x), df), where df was supplied.
}

